+++ Engine/src/SceneObjects/Sprite.cpp
@@ -28,7 +28,7 @@ void Prisma::Sprite::numSprites(unsigned int numSprites)
 {
     m_numSprites = numSprites;
     m_ssbo->resize(sizeof(glm::mat4) * m_numSprites);
-    m_ssboIds->resize(sizeof(glm::vec4) * m_numSprites);
+    m_ssboIds->resize(sizeof(glm::ivec4) * m_numSprites);
     std::vector<glm::mat4> spriteModels;
     spriteModels.resize(m_numSprites);
     glm::mat4 defaultData(1.0f);
